Move CEDAR Intake schema versions into constants file
